+++ addons/docs/src/blocks/DocsContainer.tsx
@@ -3,10 +3,10 @@
 import React from 'react';
 import { MDXProvider } from '@mdx-js/react';
 import { Global, createGlobal, ThemeProvider, ensure as ensureTheme } from '@storybook/theming';
-import { DocumentFormatting } from '@storybook/components';
+import { DocumentFormatting, DocsWrapper, DocsContent } from '@storybook/components';
 import { DocsContextProps, DocsContext } from './DocsContext';
 
-interface DocsWrapperProps {
+interface DocsContainerProps {
   context: DocsContextProps;
   content: React.ElementType<any>;
 }
@@ -26,7 +26,7 @@ const globalWithOverflow = (args: any) => {
   };
 };
 
-export const DocsWrapper: React.FunctionComponent<DocsWrapperProps> = ({
+export const DocsContainer: React.FunctionComponent<DocsContainerProps> = ({
   context,
   content: MDXContent,
 }) => {
@@ -40,9 +40,11 @@ export const DocsWrapper: React.FunctionComponent<DocsWrapperProps> = ({
       <ThemeProvider theme={theme}>
         <Global styles={globalWithOverflow} />
         <MDXProvider components={components}>
-          <DocumentFormatting>
-            <MDXContent components={components} />
-          </DocumentFormatting>
+          <DocsWrapper>
+            <DocsContent>
+              <MDXContent components={components} />
+            </DocsContent>
+          </DocsWrapper>
         </MDXProvider>
       </ThemeProvider>
     </DocsContext.Provider>