+++ addons/docs/src/blocks/index.ts
@@ -1,6 +1,8 @@
+export { ColorPalette, ColorItem, IconGallery, IconItem, Typeset } from '@storybook/components';
+
 export * from './DocsContext';
 export * from './DocsPage';
-export * from './DocsWrapper';
+export * from './DocsContainer';
 export * from './Preview';
 export * from './Props';
 export * from './Source';