+++ examples/cra-mdx/.storybook/config.js
@@ -9,11 +9,8 @@ addParameters({
     docs: {
       inlineStories: true,
       getPropDefs,
-      components: {
-        // p: ({ children }) => <b>{children}</b>,
-      },
     },
-    theme: themes.dark,
+    theme: themes.light,
   },
   docs: DocsPage,
   notes: 'global notes',