+++ examples/vue-kitchen-sink/.storybook/webpack.config.js
@@ -3,7 +3,12 @@ const path = require('path');
 module.exports = async ({ config }) => {
   config.module.rules.push({
     test: [/\.stories\.js$/, /index\.js$/],
-    loaders: [require.resolve('@storybook/addon-storysource/loader')],
+    loaders: [
+      {
+        loader: require.resolve('@storybook/addon-storysource/loader'),
+        options: { injectParameters: true },
+      },
+    ],
     include: [path.resolve(__dirname, '../src')],
     enforce: 'pre',
   });