+++ lib/components/src/blocks/DocsPage.stories.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { DocsPage } from './DocsPage';
+import * as storyStories from './Story.stories';
 import * as previewStories from './Preview.stories';
 import * as propsTableStories from './PropsTable/PropsTable.stories';
 import * as sourceStories from './Source.stories';
@@ -11,51 +12,48 @@ export const componentMeta = {
 };
 
 export const empty = () => (
-  <DocsPage
-    title={null}
-    previewProps={previewStories.error().props}
-    propsTableProps={propsTableStories.error().props}
-    sourceProps={sourceStories.sourceUnavailable().props}
-  />
+  <DocsPage title={null}>
+    {storyStories.error()}
+    {propsTableStories.error()}
+    {sourceStories.sourceUnavailable()}
+  </DocsPage>
 );
 
 export const noText = () => (
-  <DocsPage
-    title="no text"
-    previewProps={previewStories.inline().props}
-    propsTableProps={propsTableStories.normal().props}
-    sourceProps={sourceStories.jsx().props}
-  />
+  <DocsPage title="no text">
+    {previewStories.single()}
+    {propsTableStories.normal()}
+    {sourceStories.jsx()}
+  </DocsPage>
 );
 noText.title = 'no text';
 
 export const text = () => (
-  <DocsPage
-    title="Sensorium"
-    descriptionProps={descriptionStories.text().props}
-    previewProps={previewStories.inline().props}
-    propsTableProps={propsTableStories.normal().props}
-    sourceProps={sourceStories.jsx().props}
-  />
+  <DocsPage title="Sensorium">
+    {descriptionStories.text()}
+    {previewStories.single()}
+    {propsTableStories.normal()}
+    {sourceStories.jsx()}
+  </DocsPage>
 );
 
 export const withSubtitle = () => (
   <DocsPage
     title="SimStim"
     subtitle="A digital representation of the thoughts and feelings of another person."
-    descriptionProps={descriptionStories.text().props}
-    previewProps={previewStories.inline().props}
-    propsTableProps={propsTableStories.normal().props}
-    sourceProps={sourceStories.jsx().props}
-  />
+  >
+    {descriptionStories.text()}
+    {previewStories.single()}
+    {propsTableStories.normal()}
+    {sourceStories.jsx()}
+  </DocsPage>
 );
 
 export const markdown = () => (
-  <DocsPage
-    title="markdown"
-    descriptionProps={descriptionStories.markdown().props}
-    previewProps={previewStories.inline().props}
-    propsTableProps={propsTableStories.normal().props}
-    sourceProps={sourceStories.jsx().props}
-  />
+  <DocsPage title="markdown">
+    {descriptionStories.markdown()}
+    {previewStories.single()}
+    {propsTableStories.normal()}
+    {sourceStories.jsx()}
+  </DocsPage>
 );