+++ lib/components/src/blocks/DocsPage.tsx
@@ -2,14 +2,15 @@ import React from 'react';
 import { styled } from '@storybook/theming';
 import { transparentize } from 'polished';
 
-import { Preview, PreviewProps } from './Preview';
-import { PropsTable, PropsTableProps } from './PropsTable/PropsTable';
-import { Source, SourceProps } from './Source';
-import { Description, DescriptionProps } from './Description';
 import { DocumentFormatting } from '../typography/DocumentFormatting';
 
 export const breakpoint = 600;
 export const pageMargin = '5.55555';
+
+export interface DocsPageProps {
+  title: string;
+  subtitle?: string;
+}
 
 const Title = styled.h1(({ theme }) => ({
   // overrides h1 in DocumentFormatting
@@ -51,7 +52,7 @@ export const Content = styled(DocumentFormatting)({
   width: '100%',
 });
 
-const Wrapper = styled.div(({ theme }) => ({
+export const Wrapper = styled.div(({ theme }) => ({
   background: theme.background.content,
   display: 'flex',
   justifyContent: 'center',
@@ -61,31 +62,12 @@ const Wrapper = styled.div(({ theme }) => ({
   [`@media (min-width: ${breakpoint * 1}px)`]: {},
 }));
 
-export interface DocsPageProps {
-  title: string;
-  subtitle?: string;
-  descriptionProps: DescriptionProps;
-  previewProps: PreviewProps;
-  propsTableProps: PropsTableProps;
-  sourceProps: SourceProps;
-}
-
-const DocsPage: React.FunctionComponent<DocsPageProps> = ({
-  title,
-  subtitle,
-  descriptionProps,
-  previewProps,
-  propsTableProps,
-  sourceProps,
-}) => (
+const DocsPage: React.FunctionComponent<DocsPageProps> = ({ title, subtitle, children }) => (
   <Wrapper>
     <Content>
       {title && <Title>{title}</Title>}
       {subtitle && <Subtitle>{subtitle}</Subtitle>}
-      {descriptionProps && <Description {...descriptionProps} />}
-      {previewProps && <Preview {...previewProps} />}
-      {sourceProps && <Source {...sourceProps} />}
-      {propsTableProps && <PropsTable {...propsTableProps} />}
+      {children}
     </Content>
   </Wrapper>
 );