+++ lib/components/src/blocks/index.ts
@@ -1,6 +1,15 @@
 export * from './Description';
-export * from './DocsPage';
+export {
+  DocsPage,
+  DocsPageProps,
+  Wrapper as DocsWrapper,
+  Content as DocsContent,
+} from './DocsPage';
 export * from './Preview';
 export * from './PropsTable/PropsTable';
 export * from './Source';
+export * from './Story';
 export * from './IFrame';
+export * from './Typeset';
+export * from './ColorPalette';
+export * from './IconGallery';